Read fst or HDD files as DT

This is the function read_fst but with automatic conversion to data.table. It also allows to read hdd data.

readfst(path, columns = NULL, from = 1, to = NULL, confirm = FALSE)

Arguments

path

Path to fst file -- or path to hdd data. For hdd files, there is a

columns

Column names to read. The default is to read all columns. Ignored for hdd files.

from

Read data starting from this row number. Ignored for hdd files.

to

Read data up until this row number. The default is to read to the last row of the stored data set. Ignored for hdd files.

confirm

If the HDD file is larger than ten times the variable getHdd_extract.cap(), then by default an error is raised. To anyway read the data, use confirm = TRUE. You can set the data cap with the function setHdd_extract.cap, the default being 1GB.

See also

Author

Laurent Berge

Examples

# Toy example with the iris data set

# writing a hdd file
hdd_path = tempfile()
write_hdd(iris, hdd_path, rowsPerChunk = 30)

# reading the full data in memory
base_mem = readfst(hdd_path)

# is equivalent to:
base_hdd = hdd(hdd_path)
base_mem_bis = base_hdd[]