Summary information for HDD objects

Provides summary information -- i.e. dimension, size on disk, path, number of slices -- of hard drive data sets (HDD).

# S3 method for hdd
summary(object, ...)

Arguments

object

A HDD object.

...

Not currently used.

Value

This function does not return anything. It only prints general information on the data set in the console.

Details

Displays concisely general information on the HDD object: its size on disk, the number of files it is made of, its location on disk and the number of rows and columns.

Note that each HDD object contain the text file “_hdd.txt” in their folder also containing this information.

To obtain how the HDD object was constructed, use function origin.

See also

Author

Laurent Berge

Examples

# Toy example with iris data
iris_path = tempfile()
fwrite(iris, iris_path)

# destination path
hdd_path = tempfile()

# reading the text file with 50 rows chunks:
txt2hdd(iris_path, dirDest = hdd_path, rowsPerChunk = 50)

# creating a HDD object
base_hdd = hdd(hdd_path)

# Summary information on the whole data set
summary(base_hdd)
#> Hard drive data of 7.28 KB. Made of 3 files.
#> Location: C:/Users/lrberge/AppData/Local/Temp/Rtmpa0wfuK/file56886370696f/
#> 150 lines, 5 variables.

# Looking at it like a regular data.frame
print(base_hdd)
#>      Sepal.Length Sepal.Width Petal.Length Petal.Width   Species
#> 1             5.1         3.5          1.4         0.2    setosa
#> 2             4.9           3          1.4         0.2    setosa
#> 3             4.7         3.2          1.3         0.2    setosa
#> ----                                                            
#> 148           6.5           3          5.2           2 virginica
#> 149           6.2         3.4          5.4         2.3 virginica
#> 150           5.9           3          5.1         1.8 virginica
dim(base_hdd)
#> [1] 150   5
names(base_hdd)
#> [1] "Sepal.Length" "Sepal.Width"  "Petal.Length" "Petal.Width"  "Species"