Sorts HDD objects

This function sets a key to a HDD file. It creates a copy of the HDD file sorted by the key. Note that the sorting process is very time consuming.

hdd_setkey(x, key, newfile, chunkMB = 500, replace = FALSE, verbose = 1)

Arguments

x

A hdd file.

key

A character vector of the keys.

newfile

Destination of the result, i.e., a destination folder that will receive the HDD data.

chunkMB

The size of chunks used to sort the data. Default is 500MB. The bigger this number the faster the sorting is (depends on your memory available though).

replace

Default is FALSE: if the destination folder already contains data, whether to replace it.

verbose

Numeric, default is 1. Whether to display information on the advancement of the algorithm. If equal to 0, nothing is displayed.

Details

This function is provided for convenience reason: it does the job of sorting the data and ensuring consistency across files, but it is very slow since it involves copying several times the entire data set. To be used parsimoniously.

See also

Author

Laurent Berge

Examples

# Toy example with iris data

# Creating HDD data to be sorted
hdd_path = tempfile() # => folder where the data will be saved
write_hdd(iris, hdd_path)
# Let's add data to it
for(i in 1:10) write_hdd(iris, hdd_path, add = TRUE)

base_hdd = hdd(hdd_path)
summary(base_hdd)
#> Hard drive data of 48.7 KB. Made of 11 files.
#> Location: C:/Users/lrberge/AppData/Local/Temp/RtmpKImlA3/filec47846171784/
#> 1650 lines, 5 variables.

# Sorting by Sepal.Width
hdd_sorted = tempfile()
# we use a very small chunkMB to show how the function works
hdd_setkey(base_hdd, key = "Sepal.Width",
       newfile = hdd_sorted, chunkMB = 0.010)
#> Guessing R size
#> ...
#> 9 files...
#> .
#> .
#> .
#> .
#> +
#> 4
#> .
#> .
#> .
#> .
#> +
#> 4
#> .
#> .
#> .
#> .
#> +
#> 0
#> 

base_hdd_sorted = hdd(hdd_sorted)
summary(base_hdd_sorted) # => additional line "Sorted by:"
#> Hard drive data of 16.6 KB. Made of 8 files.
#> Sorted by: Sepal.Width 
#> Location: C:/Users/lrberge/AppData/Local/Temp/RtmpKImlA3/filec47831cb1ecd/
#> 1650 lines, 5 variables.
print(base_hdd_sorted)
#>       Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 1                5           2          3.5           1 versicolor
#> 2                5           2          3.5           1 versicolor
#> 3                5           2          3.5           1 versicolor
#> -----                                                             
#> 1648           5.7         4.4          1.5         0.4     setosa
#> 1649           5.7         4.4          1.5         0.4     setosa
#> 1650           5.7         4.4          1.5         0.4     setosa

# Sort with two keys:
hdd_sorted = tempfile()
# we use a very small chunkMB to show how the function works
hdd_setkey(base_hdd, key = c("Species", "Sepal.Width"),
       newfile = hdd_sorted, chunkMB = 0.010)
#> Guessing R size
#> ...
#> 9 files...
#> .
#> .
#> .
#> .
#> +
#> 4
#> .
#> .
#> .
#> .
#> +
#> 4
#> .
#> .
#> .
#> .
#> +
#> 3
#> .
#> .
#> .
#> +
#> 0
#> 

base_hdd_sorted = hdd(hdd_sorted)
summary(base_hdd_sorted)
#> Hard drive data of 16.8 KB. Made of 8 files.
#> Sorted by: Species, Sepal.Width 
#> Location: C:/Users/lrberge/AppData/Local/Temp/RtmpKImlA3/filec478132721f/
#> 1650 lines, 5 variables.
print(base_hdd_sorted)
#>       Sepal.Length Sepal.Width Petal.Length Petal.Width   Species
#> 1              4.5         2.3          1.3         0.3    setosa
#> 2              4.5         2.3          1.3         0.3    setosa
#> 3              4.5         2.3          1.3         0.3    setosa
#> -----                                                            
#> 1648           7.9         3.8          6.4           2 virginica
#> 1649           7.7         3.8          6.7         2.2 virginica
#> 1650           7.9         3.8          6.4           2 virginica